// src/app/reglage/models/campaign-model.ts
export class CampaignModel {
    id: number | undefined;
    nom: string | undefined;
    date_debut: Date | undefined;
    date_fin: Date | undefined;
    aire_sante: string | undefined;
    nbre_agent: number | undefined;
    signature: string | undefined;
    created: Date | undefined;
    update_created: Date | undefined;
}